Break config and imports into multiple files
At this point in time, EmailGUI is importing '*'.  This is not a
permanent fix and will be changed in the future, importing parts
as needed.
